+++ generator/rtcm/extract/auxiliary_information.cpp
@@ -5,6 +5,7 @@
 #pragma GCC diagnostic ignored "-Wreserved-macro-identifier"
 #pragma GCC diagnostic ignored "-Wreserved-identifier"
 #pragma GCC diagnostic ignored "-Wundef"
+#pragma GCC diagnostic ignored "-Wunused-function"
 #include <GNSS-AuxiliaryInformation.h>
 #include <GNSS-ID-GLONASS-SatElement.h>
 #include <GNSS-ID-GLONASS.h>